move code, fix test case creation, fix some server tests
