+++ src/test/java/de/tum/in/www1/artemis/hestia/ExerciseHintIntegrationTest.java
@@ -469,7 +469,7 @@ class ExerciseHintIntegrationTest extends AbstractSpringIntegrationBambooBitbuck
         for (ProgrammingExerciseTestCase testCase : exercise.getTestCases()) {
             var feedback = new Feedback();
             feedback.setPositive(successfulTestCases.contains(testCase));
-            feedback.setText(testCase.getTestName());
+            feedback.setTestCase(testCase);
             feedback.setVisibility(Visibility.ALWAYS);
             feedback.setType(FeedbackType.AUTOMATIC);
             database.addFeedbackToResult(feedback, result);
