+++ src/test/java/de/tum/in/www1/artemis/hestia/ExerciseHintServiceTest.java
@@ -331,7 +331,7 @@ class ExerciseHintServiceTest extends AbstractSpringIntegrationBambooBitbucketJi
         for (ProgrammingExerciseTestCase testCase : exercise.getTestCases()) {
             var feedback = new Feedback();
             feedback.setPositive(successfulTestCases.contains(testCase));
-            feedback.setText(testCase.getTestName());
+            feedback.setTestCase(testCase);
             feedback.setVisibility(Visibility.ALWAYS);
             feedback.setType(FeedbackType.AUTOMATIC);
             database.addFeedbackToResult(feedback, result);
